refactor: CapturedVariableProvider tests merged into one file
